removed the old group/section code and flattened the backend file structure
