+++ crates/pacdef/src/backend/dnf.rs
@@ -2,7 +2,7 @@ use std::collections::BTreeMap;
 
 use anyhow::Result;
 
-use crate::backend::root::{run_args, run_args_for_stdout};
+use crate::backend::{run_args, run_args_for_stdout};
 use crate::prelude::*;
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
@@ -12,6 +12,7 @@ pub struct DnfQueryInfo {
     user: bool,
 }
 
+#[derive(Debug, Clone)]
 pub struct DnfInstallOptions {
     repo: Option<String>,
 }
