+++ crates/pacdef/src/backend/pipx.rs
@@ -5,43 +5,25 @@ use anyhow::Context;
 use anyhow::Result;
 use serde_json::Value;
 
-use crate::backend::root::run_args;
-use crate::backend::root::run_args_for_stdout;
+use crate::backend::run_args;
+use crate::backend::run_args_for_stdout;
 use crate::prelude::*;
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub struct Pip;
+pub struct Pipx;
 
-pub struct PipQueryInfo {
-    explicit: bool,
-}
-
-impl Backend for Pip {
+impl Backend for Pipx {
     type PackageId = String;
     type RemoveOptions = ();
     type InstallOptions = ();
-    type QueryInfo = PipQueryInfo;
+    type QueryInfo = ();
     type Modification = ();
 
     fn query_installed_packages(_: &Config) -> Result<BTreeMap<Self::PackageId, Self::QueryInfo>> {
-        let all = extract_package_names(run_args_for_stdout(
-            ["pip", "list", "--format", "json"].into_iter(),
-        )?)?;
-        let implicit = extract_package_names(run_args_for_stdout(
-            ["pip", "list", "--format", "json", "--not-required"].into_iter(),
-        )?)?;
+        let names =
+            extract_package_names(run_args_for_stdout(["pipx", "list", "--json"].into_iter())?)?;
 
-        let explicit = all.difference(&implicit);
-
-        Ok(implicit
-            .into_iter()
-            .map(|x| (x, PipQueryInfo { explicit: false }))
-            .chain(
-                explicit
-                    .into_iter()
-                    .map(|x| (x.clone(), PipQueryInfo { explicit: true })),
-            )
-            .collect())
+        Ok(names.into_iter().map(|x| (x, ())).collect())
     }
 
     fn install_packages(
@@ -50,7 +32,7 @@ impl Backend for Pip {
         _: &Config,
     ) -> Result<()> {
         run_args(
-            ["pip", "install"]
+            ["pipx", "install"]
                 .into_iter()
                 .chain(packages.keys().map(String::as_str)),
         )
@@ -69,7 +51,7 @@ impl Backend for Pip {
         _: &Config,
     ) -> Result<()> {
         run_args(
-            ["pip", "uninstall"]
+            ["pipx", "uninstall"]
                 .into_iter()
                 .chain(packages.keys().map(String::as_str)),
         )
@@ -79,23 +61,12 @@ impl Backend for Pip {
 fn extract_package_names(stdout: String) -> Result<BTreeSet<String>> {
     let value: Value = serde_json::from_str(&stdout)?;
 
-    let result = value
-        .as_array()
-        .context("getting inner json array")?
-        .iter()
-        .map(|node| node["name"].as_str().expect("should always be a string"))
-        .map(String::from)
-        .collect();
-
-    Ok(result)
-}
-
-fn extract_pacdef_packages_pipx(value: Value) -> Result<Packages> {
     let result = value["venvs"]
         .as_object()
         .context("getting inner json object")?
         .iter()
-        .map(|(name, _)| Package::from(name.as_str()))
+        .map(|(name, _)| name.clone())
         .collect();
+
     Ok(result)
 }
