+++ crates/pacdef/src/backend/rustup.rs
@@ -1,5 +1,5 @@
-use crate::backend::root::run_args;
-use crate::backend::root::run_args_for_stdout;
+use crate::backend::run_args;
+use crate::backend::run_args_for_stdout;
 use crate::prelude::*;
 use anyhow::Result;
 use std::collections::BTreeMap;
